Remove manual complexity check
With a working code-climate integration, we now have more detailed,
legible and specific reports on code maintainability. These comments are
available in the PR checks and they account for the diff - giving insights
relevant to the files and changes in that PR. Hence the old radon based 
complexity analysis as part of the Github Actions is redundant, hence removed.