Drop support for perfect simulator
The perfect simulator allows a lightweight simulation that
doesn't calculate the full quantum dynamics but uses ideal
gate matrices. Support for this simulator is being dropped
since it requires a rework to either be merged with the physics
simulator (with a dry_run argument) or be rewritten to include
useful tools such as classical noise modelling etc.